test: refactor parser unit tests into modules
